backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0
email-validator>=2.1.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
python-docx>=1.1.0
pypdf2>=3.0.1
pdfplumber>=0.10.3
google-generativeai>=0.3.2
langchain>=0.1.0
langchain-google-genai>=0.0.6
boto3>=1.34.0
httpx>=0.26.0
tenacity>=8.2.3
python-dotenv>=1.0.0
aiofiles>=23.2.1